MultiKernelBench/requirements.txt
numpy
openai
packaging
torch==2.1.0
numpy==1.26.4
decorator
scipy
psutil
google
tvm
protobuf
tqdm
datasets
transformers
together
pytest
ninja
einops
pyyaml
setuptools